added support for sync, and async features at the same time
