+++ src/lib.rs
@@ -1,4 +1,4 @@
-#![deny(missing_docs)]
+#![warn(missing_docs)]
 //! A pure-Rust lowlevel library for controlling wpasupplicant remotely
 //!
 //! Note that in order to connect to wpasupplicant, you may need
@@ -16,6 +16,10 @@
 
 mod error;
 mod wpactrl;
+/// enables syncronous operation of this crate
+#[cfg(feature = "sync")]
+pub mod sync;
+//pub mod hostapd;
 pub use crate::wpactrl::{Client, ClientAttached, ClientBuilder};
 
 pub use crate::error::Error;
