+++ station.txt
@@ -1,0 +1,87 @@
+stas: 56:b3:48:c1:22:7e
+flags=[AUTH][ASSOC][AUTHORIZED][WMM][HT][VHT]
+aid=1
+capability=0x1411
+listen_interval=10
+supported_rates=02 04 0b 16 0c 12 18 24 30 48 60 6c
+timeout_next=NULLFUNC POLL
+dot11RSNAStatsSTAAddress=56:b3:48:c1:22:7e
+dot11RSNAStatsVersion=1
+dot11RSNAStatsSelectedPairwiseCipher=00-0f-ac-4
+dot11RSNAStatsTKIPLocalMICFailures=0
+dot11RSNAStatsTKIPRemoteMICFailures=0
+wpa=2
+AKMSuiteSelector=00-0f-ac-1
+hostapdWPAPTKState=11
+hostapdWPAPTKGroupState=0
+dot1xPaePortNumber=1
+dot1xPaePortProtocolVersion=2
+dot1xPaePortCapabilities=1
+dot1xPaePortInitialize=0
+dot1xPaePortReauthenticate=FALSE
+dot1xAuthPaeState=5
+dot1xAuthBackendAuthState=6
+dot1xAuthAdminControlledDirections=0
+dot1xAuthOperControlledDirections=0
+dot1xAuthAuthControlledPortStatus=1
+dot1xAuthAuthControlledPortControl=2
+dot1xAuthQuietPeriod=60
+dot1xAuthServerTimeout=30
+dot1xAuthReAuthPeriod=3600
+dot1xAuthReAuthEnabled=TRUE
+dot1xAuthKeyTxEnabled=FALSE
+dot1xAuthEapolFramesRx=12
+dot1xAuthEapolFramesTx=13
+dot1xAuthEapolStartFramesRx=0
+dot1xAuthEapolLogoffFramesRx=0
+dot1xAuthEapolRespIdFramesRx=1
+dot1xAuthEapolRespFramesRx=10
+dot1xAuthEapolReqIdFramesTx=1
+dot1xAuthEapolReqFramesTx=10
+dot1xAuthInvalidEapolFramesRx=0
+dot1xAuthEapLengthErrorFramesRx=0
+dot1xAuthLastEapolFrameVersion=1
+dot1xAuthLastEapolFrameSource=56:b3:48:c1:22:7e
+dot1xAuthEntersConnecting=1
+dot1xAuthEapLogoffsWhileConnecting=0
+dot1xAuthEntersAuthenticating=0
+dot1xAuthAuthSuccessesWhileAuthenticating=1
+dot1xAuthAuthTimeoutsWhileAuthenticating=0
+dot1xAuthAuthFailWhileAuthenticating=0
+dot1xAuthAuthEapStartsWhileAuthenticating=0
+dot1xAuthAuthEapLogoffWhileAuthenticating=0
+dot1xAuthAuthReauthsWhileAuthenticated=0
+dot1xAuthAuthEapStartsWhileAuthenticated=0
+dot1xAuthAuthEapLogoffWhileAuthenticated=0
+dot1xAuthBackendResponses=10
+dot1xAuthBackendAccessChallenges=9
+dot1xAuthBackendOtherRequestsToSupplicant=10
+dot1xAuthBackendAuthSuccesses=1
+dot1xAuthBackendAuthFails=0
+dot1xAuthSessionId=14146D481E8A487B
+dot1xAuthSessionAuthenticMethod=1
+dot1xAuthSessionTime=958
+dot1xAuthSessionTerminateCause=999
+dot1xAuthSessionUserName=lazaras@pdx.edu
+authMultiSessionId=CB233FF911713422
+last_eap_type_as=25 (PEAP)
+last_eap_type_sta=25 (PEAP)
+rx_packets=1467
+tx_packets=523
+rx_bytes=132714
+tx_bytes=291292
+inactive_msec=1043
+signal=-25
+rx_rate_info=10
+tx_rate_info=650 mcs 7
+rx_vht_mcs_map=fffe
+tx_vht_mcs_map=fffe
+ht_mcs_bitmask=ff000000000000000000
+connected_time=958
+supp_op_classes=51515354737475767778797a7b7c7d7e7f808182
+min_txpower=8
+max_txpower=20
+vht_caps_info=0x03800100
+ht_caps_info=0x013c
+ext_capab=00000a8200400040000120
+
